pyproject.toml
[tool.poetry]
name = "postgres-ai-agent"
version = "0.1.0"
description = ""
authors = ["JamieCattoCode <jamietcatto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.18"
psycopg2-binary = "^2.9.9"
argparse = "^1.4.0"
python-dotenv = "^1.0.0"
openai = "0.27.6"
aider-chat = "^0.16.3"
pyautogen = "^0.1.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "postgres_ai_agent.main:main"